requirements.txt
Flask==3.0.0
google-generativeai==0.4.0
pinecone-client==2.2.4
langchain==0.1.16
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.1
gunicorn==21.2.0
